Separate tests by functonality.
Facilitate team development of tests.
